Change CORS headers to be generated via function
This is done primarily because we need to allow the allowed origin to
come from one of a few places. In the future, we may decide to have this
as an environment variable, but for now we need to be able to verify the
origin actually comes from a pool of authorized origins.